refactor(event): views/crud.py を Event 用 / EventDetail 用に分割
623行に2モデル×6 CBV が同居していた。境界 (467/468行) で crud_event.py (463行) と
crud_event_detail.py (174行) に分割。ロジックは無変更の移動のみ。
logger 名が event.views.crud → crud_event / crud_event_detail に変わる点のみ注意
（Sentry フィルタは logger 名非依存で送信挙動不変）。

検証: event 全テスト緑（時刻依存の既存フレーク2件を除く）
